requirements.txt
arabic-reshaper==2.1.1
asgiref==3.2.10
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.14.2
cssselect2==0.3.0
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1
django-extensions==3.0.7
django-import-export==2.3.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
jdcal==1.4.1
lxml==4.5.2
MarkupPy==1.14
numpy==1.19.1
odfpy==1.4.1
openpyxl==3.0.5
pandas==1.1.1
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
PyPDF2==1.26.0
Pyphen==0.9.5
python-bidi==0.4.2
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.1
PyYAML==5.3.1
reportlab==3.5.59
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
tinycss2==1.0.2
unicodecsv==0.14.1
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
xlrd==1.2.0
xlwt==1.3.0
